chore: Refactor the leader election
