+++ workflow/controller/operator_test.go
@@ -3565,7 +3565,7 @@ spec:
 }
 
 func getEvents(controller *WorkflowController, num int) []string {
-	c := controller.eventRecorderManager.(*testEventRecorderManager).eventRecorder.Events
+	c := controller.EventRecorderManager.(*testEventRecorderManager).eventRecorder.Events
 	events := make([]string, num)
 	for i := 0; i < num; i++ {
 		events[i] = <-c
